Move AccessibleContainerView to its own file
